Move compiler to its own package.
`Compile()` now returns the obj instead of VM, to remove a circular dependency as
`vm.Loader.CompileAndRun()` calls `Compile()`.

We also get to remove some parameters as they're only needed for constructing
the VM.
