+++ internal/vm/compiler/compiler_test.go
@@ -1,18 +1,18 @@
 // Copyright 2019 Google Inc. All Rights Reserved.
 // This file is available under the Apache license.
 
-package vm_test
+package compiler_test
 
 import (
 	"strings"
 	"testing"
 
-	"github.com/google/mtail/internal/vm"
+	"github.com/google/mtail/internal/vm/compiler"
 )
 
 func TestCompileParserError(t *testing.T) {
 	r := strings.NewReader("bad program")
-	_, err := vm.Compile("test", r, true, true, true, nil, 0, 0, false)
+	_, err := compiler.Compile("test", r, true, true, 0, 0)
 	if err == nil {
 		t.Errorf("expected error, got nil")
 	}
@@ -22,7 +22,7 @@ func TestCompileCheckerError(t *testing.T) {
 	r := strings.NewReader(`// {
 i++
 }`)
-	_, err := vm.Compile("test", r, true, true, true, nil, 0, 0, false)
+	_, err := compiler.Compile("test", r, true, true, 0, 0)
 	if err == nil {
 		t.Error("expected error, got nil")
 	}
@@ -33,7 +33,7 @@ func TestCompileCodegen(t *testing.T) {
 // {
   i++
 }`)
-	_, err := vm.Compile("test", r, true, true, true, nil, 0, 0, false)
+	_, err := compiler.Compile("test", r, true, true, 0, 0)
 	if err != nil {
 		t.Error(err)
 	}
